requirements.txt
dj-config-url==0.1.1
Django==2.0.1
gitdb2==2.0.3
GitPython==2.1.8
pytz==2017.3
smmap2==2.0.3
